requirements.txt
onnx<=1.6.0  # nart 0.2.4 incompatible with onnx 1.7.0
prettytable
scipy==1.1.0
coverage==5.0.3
pycocotools-fix
pandas>=0.25.1
easydict>=1.9
opencv-python>=4.1.0
Cython>=0.29.2,<0.29.15
tensorboardX==1.8.0
matplotlib==3.3.3
pillow==6.2.1
lvis
SharedArray
numba==0.48.0
psutil==5.9.0
scikit-learn==0.21.3
open3d==0.10.0
onnxruntime